+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+  - Added logger debug statements for debugging.
+
 ## [1.0.2] - 2018-02-08
 
 ### Changed